Do not use mode: preserve
The linter is wrong.
It is better to create files with the default umask.

mode: preserve always set 400, because git does not track all
permissions - it only tracks the executable flag.

This can cause some issues, like in this instance makepkg not being able
to read /etc/pacman.d/mirrorlist