fzf: Update search with rg directly
Otherwise we just do the initial search with ripgrep and after that
filter with fzf, which has some disadvantages.

For example one problem was that starting with the empty query would
return all lines, then filtering with fzf would return wrong results
because it would search in the returned filenames as well (as opposed to
file contents), even when narrowing the search.