Use alternative notmuch-mutt integration
This has the advantage that it does not use the hack of changing to a
virtual mailbox. Instead it uses the native limit functionality from
mutt.

For example this allows us to actually delete mails while we limit with
notmuch (not possible with virtual mailboxes) and it also integrates
better with the rest of the config, for example FCC-copying sent mails
in the correct mailbox now works by default again.

The upstream example limits search results to 600 with "head 600", let's
see if we should better remove that in the future.

The only feature missing right now, is that typing '?' no longer shows
the notmuch documentation, but we could easily implement that ourself
if we really need it.

refs: https://notmuchmail.org/notmuch-mutt/
refs: http://log.or.cz/?p=228