Add systray icon to mail sync
For now this depends on yad, but in the future we can probably implement
this in raw shell.

For now this doesn't show any tooltip, in --listen mode we should be
able to provide the tooltip text with "tooltip:STRING", but it doesn't
seem to work, we will have to investigate this further.

Nonetheless we still start in --listen mode, because this causes the
default left-click action to no longer kill the process.

Right now we should not allow the systray to be killed, because from the
script we later call "kill" on the PID of that process, which would be
susceptible to a TOCTOU (pid reuse for another process), if we allowed
it to be killed.

It can still be killed with a middle-click, but for now let's assume
that the user would never do this .