Use FQDN of hostname
/etc/hosts still needs only the host name without domain, so we strip
it.
But in /etc/hosts we can configure the short hostname to be an alias for
the hostname with FQDN, so we do that.

This will fix mutt generating message IDs that contain localdomain,
which is a bit cringe on mailing lists.

Instead we can now use our own domain, which will then automatically be
used by mutt for generating message IDs.