Fix get-pkgbuild for new git upstream
Recently Arch Linux migrated the package repositories to git [0].
As a result of this, the asp tool was dropped.

Theoretically, a replacement tool called pkgctl is available in the
devtools package, but it seems they are still bikeshedding about how to
enable cloning for unprivileged users [1]. Right now, it clones with ssh
by default, which means it will fail for normal users.

To fix this, just clone manually over https. In order to force the
command to fail if the package is not in the official repository, we
disable git asking for a password for possibly hidden repos by setting
GIT_TERMINAL_PROMPT to false.

[0] https://archlinux.org/news/git-migration-completed/
[1] https://gitlab.archlinux.org/archlinux/devtools/-/issues/132