Fix macro shortcut
The intention of Q was to "just work" and replay the last replayed
macro, but this doesn't work if there never was a macro replayed to
begin with.

Fix this by just always using the q register: We are almost always using
it anyway and it makes sense for the Q key to replay the q register.