Add tmux binding to search down from the last prompt
This first navigates to the very beginning of the last prompt using OSC
133 [0] and then searches down from that position.

This can be further improved in the future, e.g. it could navigate to
the beginning of the output (and exclude the prompt), but this requires
further upstream tmux changes.

Note that I also tried merging the two "previous-prompt" commands into
one by using "-N 2" on "send-keys", but it didn't work for some reason.

[0] https://github.com/tmux/tmux/pull/3596