Detect number of spaces for autoindent
This is a character optimized version of a four-way if condition.  n is
calculated as the sum of three different components, where only ever
exactly one component of the three is non-zero:

get(a:, 1, 0): The non-zero explicitly given count

get(filter(...)) * !a:0 -> If no count is given, this gets the first
number of consecutive spaces within the file. If there are no spaces, it
returns 4 (our preferred tab shiftwidth).

!get(a:, 1, 0) * a:0: If the explicitly given count is zero, this
evaluates to 4 (our preferred tab shiftwidth)

Why did I do this? I don't know, but the four-way if condition felt too
inelegant.