Simplify colored man pages
The colored-man-pages plugin recently stopped working. I couldn't figure
out exactly why, but instead found out that there is no need to use a
bloated plugin for this, as this can be achieved with a less one-liner,
as described here at the bottom [0].

Even more so, we can safely use this for all less pages, as all this
does is change bold text to red bold text, and underlined text to blue
underlined text.
And just with this little snippet we achieve everything that the
colored-man-pages in much more lines of code. :)

The --use-color argument separately activates some colored goodies in
less, such as colored search.

There also seems to be some work going on so that the MANROFFOPT
environment variable isn't necessary anymore, we should follow that [1].

[0] https://felipec.wordpress.com/2021/06/05/adventures-with-man-color/
[1] https://bugs.archlinux.org/task/71136