zgen -> zgenom
We have been using zgen since 9ec8b400848af02c597d8af81bddd34a7b351ad3
in 2018, but in the meantime zgen has become unmaintained.

Let's switch to zgenom which seems to be its successor with some extra
features, such as compiling the user zshrc (which we are activating with
this patch as well) and support for the ZSH plugin standard [0].

Note that the ~/.zgen directory needs to be manually deleted, otherwise
zgenom will keep using it for compatibility.

I am also side-eyeing zinit [1] at the moment, which seems to have
similar functionality and is of particularly high quality like
everything from psprint/zdharma.

[0] https://zdharma-continuum.github.io/Zsh-100-Commits-Club/Zsh-Plugin-Standard.html
[1] https://github.com/zdharma-continuum/zinit