Fallback to master branch if origin/HEAD doesn't exist
Previously we assumed that the origin/HEAD ref must exist, but this is
sometimes not the case. Often it does exist, because "git clone" creates
this ref, but there are many scenarios where it is not created, e.g.
when the repo is initialized locally.

In that case just fall back to master.
We could use "git config init.defaultBranch" instead, but "git config"
doesn't return values if a config option isn't set yet, so that doesn't
really help us much.

We could also use many of the commands querying the remote (e.g. git
remote show origin), but that would require an internet connection and
has much more latency.

ref: https://stackoverflow.com/a/66895556