gerrit: Prevent direct git pushes
Most Gerrit documentations recommend to modify the origin push url to
something invalid, e.g.:
git remote set-url --push origin no_push

I don't like this workflow, because it interferes with some scripts and
aliases and let's be honest, it's just a hacky workaround, no_push is
not a special set-url flag, it's just an invalid url.

Instead I discovered that some people use pre-push hooks to accomplish
the same behaviour [0], which is imho a much more elegant solution.

My implementation is essentially a pure bash rewrite of the above
solution.

All refs are passed via stdin to pre-push hooks, so we just need to grep
for the standard gerrit review refs.

[0] https://www.dvratil.cz/2017/06/git-trick-481-prevent-accidentally-pushing-into-git-instead-of-gerrit/