vim: Use internal diff
In 44206e99053148bdac595dccefc62909f1d2a875 we introduced the patience
algorithm for vim's diff, but it did not actually take effect, because
we also need to set the option to use the vim internal diff.

We should also consider if we instead want to set diffexpr, perhaps vim
can make use of some even nicer diff options.