mutt: Remove hostname from Message-ID
By default, mutt's message ids are pretty whack because it always leaks
the hostname [0], e.g.:
Message-ID: <GYGZXHFzRtV12AXC@hostname.example.com>

With this patch the message-id now looks like:
Message-ID: <20240203A140815c078047c.john@example.com>

It is indeed true that some spam filters increase the spam score if the
right hand side of the message id does not end with a domain. Therefore
we still keep the FQDN, but drop the hostname.

There is another problem with the default message id generator, which is
that the left part is base64 encoded and thus may contain a '/', which
can cause problems with LKML links where the message id is used as
permalink [1].

Luckily "%x" exists which uses hex encoding, with the drawback that it
encodes only 1 byte of random data with 2 characters. To make sure that
the entropy is high enough to avoid message id collisions, we prepend 4
bytes of random data with the timestamp down to second precision, as
well as adding the "%c" step counter.
Thus, to get a message id collision, one would have to do one of the
following:
- In a single second, open different mutt instances and starting with
  the second email sent in that second beat the chance of 1:4294967296
  (2^(8*4))
- In a single second, stay in the same mutt instance and starting with
  the second email sent in that second beat the chance of 1:111669149696
  (26*2*(8*4))

I will never be able to send multiple non-automated emails in the same
second, so in that area there never was danger anyway.

For automated emails, it is possible to send multiple emails in the same
second, but with the above the chance for message id collision should be
slim enough to never be a concern. Most of my automated messages do not
go over mutt anyway, but instead use some other mechanism (e.g. git
send-email) and go directly over msmtp.

Finally one might wonder why I do all this effort to drop the $HOST,
just to leak $USER instead. The reasoning is that it improves
readability of LKML links, e.g. most maintainers also have their name in
message ids [2]. The "kees@kernel.org" allows to immediately see who was
the author of the mail behind that link. It is fine to drop the
".$USER", but then one should make sure that entropy is high enough to
avoid message id collisions for users on the same domain.

The same readability reason is also why we have the date at the
beginning: This allows us to see at a glance when the email was sent
just from the link alone.

Implementation wise we unfortunately have to resort to format filters
and drop into a shell, the mutt-internal message id formatters are not
powerful enough.

Also something weird with the implementation is that we have to escape
"\" 4 times (so "\\\\") instead of the expected "\\". The shell should
not be at fault here as the "\" is in a quote; perhaps mutt does
expanding twice.

Note that using this message id generator probably only makes sense, if
you have your domain configured. If "hostname -f" resolves to
localdomain, this probably does not make much sense.

[0] https://gitlab.com/muttmua/mutt/-/issues/453
[1] https://people.kernel.org/monsieuricon/fix-your-mutt
[2] https://lore.kernel.org/lkml/20240130220218.it.154-kees@kernel.org/