tmux: Show number of sessions
This shows the current session id and the number of tmux server
sessions, if there are multiple, which is helpful to know if quitting
will quit tmux or switch to another running session (with the
detach-on-destroy option set).

The required tmux format string is quite cursed and this unfortunately
also means disabling tpipeline's autoembed, as the only option in tmux
to put something in front of a centred window list is to add an
"#[align=absolute-centre]" formatted string in status-left. There is no
status-centre, and putting it in status-right would mean appending to
the centred window list.

Another alternative would be to use window-status-format, but that would
be even more cursed, because it would involve conditional formatting to
check if the current window is the first window (for prepending
purposes), plus everything needs duplicating in
window-status-current-format, in case the current window is the first
window.

For the implementation, we use "" string formatting for status-left, as
this makes Unicode characters more accessible. The disadvantage is that
we need to escape backslashes once more compared to the '' string variant.
The first backslash in '\#{session_id}' exists, because the session id
expands to e.g. '$0', which we need to escape for the outer shell call
(to cat).