tmux: Improve style of session indicator
Previously it displayed the current session ID and the total number of
sessions, but that could lead to some weird displays, as session IDs may
be larger than the number of sessions, when some sessions have been
closed, e.g.: "8/3".

Now it uses a stylised Unicode dot representation for every session. The
current session is displayed with an accent color. This improves UX a
lot, as it quickly visualizes the position of the session in the list
and it usually does not cost much more space, as sessions are used
sparingly.

The only problem with the implementation is that the "#{S:}" session
iteration format variable does not differentiate between current and
other sessions, unlike the window and pane iteration formats.

Luckily it seems that both the session_id and session_name variables are
accidentally (not sure if that is an implementation detail or intended
API) iterated as well. There is no way to retrieve the current session
id inside the loop, but there is client_session to retrieve the current
session name, so we can string compare it with the iterated name
variable to know when we reach the current session.

For performance reasons this is all still hidden behind a check if the
number of sessions is greater than one.