link.sh: Do not attempt to create directories over dead symlinks
If $DEST exists, but is a symlink pointing to a missing file, the "-e"
check will fail as it dereferences symlinks and the pointed to file is
not existing.
To fix this, add an extra check for if there is a symlink with "-h",
which does not dereference the symlink.

This will avoid an error message about being unable to create a
directory where there is still a dangling symlink.

The dangling symlink would get fixed in the following lines anyway.