tmux: Allow to toggle prefix
This is very useful for nested tmux sessions (e.g. over ssh), where it
is a giant PITA to press the prefix twice to control the inner tmux
session.

So now we can instead use "prefix+F12" to toggle the prefix between ^A
and ^B, so that different prefixes can be used for both sessions.

A different approach would be to toggle all keybindings [0], but for my
case this much simpler toggle is sufficient, as we do not have any
keybindings without the prefix.

We could also autodetect if in ssh, but I prefer to manually trigger
these nested tmux session remappings to avoid confusions.

[0] https://github.com/tmux/tmux/issues/237