Avoid $MAIL
This causes some very weird errors when the mail is very long (more than
1MB):

/usr/bin/jq: Argument list too long

The best guess is that notmuch tries to use the $MAIL variable to guess
where MAIL is, but we then overwrite the location to the content.

To fix this, we avoid touching this reserved variable, which usually
points to a location in /var/spool.