Allow to automatically push all branches in branchless mode
This does not push branches if they do not have a remote tracking
reference yet (i.e. for the first push, each branch must still be
manually pushed) and it also does not try to push them, if the
remote-tracking reference is already up-to-date.

The latter one is less important as git-push already detects a no-op,
but it is still useful to have, e.g. when dealing with spotty internet
connections.

This operation obviously also only considers commits between the last
branch-off point and the current HEAD.