git: Auto-stash on rebase
I have been contemplating adding this option for a long time due to an
unreasonable fear, that it will allegedly forget to unstash if a rebase
with merge conflicts is aborted.

A quick test confirms that this is not the case and apparently the stash
entry is not even visible in git stash, as it is implemented as a
special temporary stash, so the concerns were unfounded after all.

This is the last snippet I am stealing from the popular Gitbutler post
[0], because the rest of them are nowhere any good.

[0] https://blog.gitbutler.com/how-git-core-devs-configure-git/