aliases-gen: Always reuse last timestamp
The notmuch query for retrieving the mails since the last sync was
cleverly chosen to also include the last already processed mail, since
we only initialize $TIMESTAMP inside the for loop, so this was a great
way to make sure we also write a valid timestamp if there are no new
mails since the last sync.

This unfortunately causes a bug when that last mail is deleted, which
will lead to $TIMESTAMP never being initialized and thus an empty
timestamp being written to the cache, which will blow up on the next run
and cause a full reindexing of all mail.

To fix this, always initialize $TIMESTAMP from the cache, which will
avoid the problem with deleted mails, there was no use for the extra
$LAST_TIMESTAMP variable in the first place.