Add arcan keyboard config
This is the beginning of the transition to using Arcan as my daily
driver (with some minor dogfooding to be done still).
For now durden with tiling set to BSP is a reasonable alternative to
bspwm, but only time will tell if I have to start writing my own WM.
Now onwards to enjoying the Wayland circus act from a safe distance.