Use English locale
The main motivation for this is that I always want to read gcc compiler
errors in English. Trying to poke at one of the dozens different LC
variables to no avail, the only remaining option is to change the main
language to English, then all LC variables back to German to avoid the
ridiculous inventions on the other side of the pond such as the imperial
system.

This obviously means that the entire system will now default to English
language - not necessarily something I intended to do, but there are
some upsides, e.g. Youtube/Reddit shitholes no longer aggressively
pushing their AI translations (recently even automatically translated
audio tracks with about the inflection quality, you would expect).

Sometimes it can be a little awkward to have basic OS interfaces appear
not in mother tongue, but at this point the benefits seem to overweight.

Regarding the selection of LC variables, we basically set the selection
that KDE sets in its language settings plus LC_CTYPE, to also overwrite
the LC_CTYPE set from arcterm (see setenv in arcterm.c).