Show notification on mute
This is definitely up there among the most cursed shell one-liners I
have written.

Every sane person would have used at least a shell script to get
variables, but we can emulate that with multi-target xargs. For the mic
mute for example we want to branch into showing the Unicode codepoints
"\UF130" and "\UF131" , so we use the Bash feature for escaping Unicode
directly (contrary to common misconception "\x" expansion happens only
at the printf/echo level), e.g. $'\UF130', but the last digit is only
known at runtime and the Unicode expansion happens at the wrong level
too, so we have to indirect call it with an extra shell.
The same is of course also true for multi-target xargs, as the expansion
happens too early, so we have to launch two sub-shells.
This does not even mention the cursed multi-escape string quotes
necessary, but at least this shit still parses correctly inside sxhkd,
which is honestly pretty impressive given that it is not meant to be
completely Bash compatible.