Allow to yank to named registers
The previous mapping for sync to the global clipboard was used instead
of set clipboard=unnamedplus, because we do not want delete operations
such as dd to copy into the global clipboard, just because they yank
into vim registers.
However this mapping completely disabled yanking to custom registers,
e.g. it was no longer possible to use "ayw to yank into register a.

The new mapping now only defaults to the clipboard register, if the
unnamed register would be used instead and only for yank and paste
operations.

This should finally make the vim clipboard work as expected without
giving up on other vim registers.