Generate geolocation info
The geoclue service tends to break regularly [0] and has questionable
dependencies [1], but this is all unnecessary in the first place:
We already know our timezone info, since every install symlinks
/etc/localtime to a timezone. The offline timezone description zone tab
file contains coordinates for all such timezones.

There is some annoying conversion to be done between the ISO-6709
sign-degrees-minutes-seconds format and decimal degree coordinates, but
in general using this offline information is a much better approach, as
it's completely unnecessary to send a network request to some shady
service just to locate our device.

The information from the timezone is a little coarser (since the
timezone city is probably not your exact city), but that is not only
okay but even appreciated for sensitive information such as the
geolocation, and it's not like network-based geolocation is super
precise either.

For all intents and purposes this coarse information is still accurate
enough in the end usecase (e.g. approximate day-night time switching).

We use the file zone1970.tab instead of zone.tab or zonenow.tab, because
it is the more complete source according to theory.html [2] from the
tzdata source.

Note that there is also a go project called geoclue-tz [3] to perform
the same thing as this script, including monitoring for timezone changes
and regenerating the geolocation, but personally I prefer this simple
script as the timezone almost never changes.
If the timezone changes, the script can just be run again.

Finally, we also need to make sure to start the geoclue agent so that
darkman can connect to retrieve the geolocation.

[0] https://github.com/NixOS/nixpkgs/issues/321121
[1] https://github.com/NixOS/nixpkgs/issues/329522
[2] https://data.iana.org/time-zones/tzdb-2022b/theory.html
[3] https://github.com/zquestz/geoclue-tz